updated swarm-admin version
The swarm admin service is now updated to 1.0.1 to also support volumes. It also
means that the installation procedure has changed slightly as is documented in
the README.md.

In short instead of ./script/edit-hosts.sh which was the script that needed to
be called originally you now have to call ./scripts/install.sh which in turn
calls the edit-hosts script but also sets the volumes for the swarm admin
correctly.